Completed assignment

+++ FVCalc.java
@@ -2,7 +2,7 @@
 public class FVCalc {
 	public static void main(String[] args){
 		int currValue = Integer.parseInt(args[0]);
-		int interestRate = Integer.parseInt(args[1]);
+		double interestRate = Double.parseDouble(args[1]);
 		int years = Integer.parseInt(args[2]);
 		int futureValue = (int) (currValue * Math.pow(1 + interestRate / 100.0, years));
 		System.out.println("After " + years + " years, $" + currValue + " saved at " + interestRate + "% will yield $" + futureValue);
